update trans2 payment

+++ common/types.go
@@ -14,10 +14,10 @@ type CreateOrderResp struct {
 }
 
 type CreateNewPaymentReq struct {
-	WarehouseID int32 `json:"c_w_id"`
-	DistrictID  int32 `json:"c_d_id"`
-	CustomerID  int32 `json:"c_id"`
-	Payment     int32 `json:"payment"`
+	WarehouseID int32   `json:"c_w_id"`
+	DistrictID  int32   `json:"c_d_id"`
+	CustomerID  int32   `json:"c_id"`
+	Payment     float64 `json:"payment"`
 }
 
 type CreateNewPaymentResp struct {
